Day 8 solution in C++
I used brute-force for part 2 as it was quite practical on the small
amount of input data. A more efficient approach would have required
building a graph of instructions and jumps, find all instructions
eventually leading to the end of the program (target addresses),
find all instructions that, if swapped, could lead to one of the target
addresses (candidate addresses), and finally find all instructions among
the candidate addresses that could be reached from address 0. There
should be only one of them in the example.